telegram-bot-service/requirements.txt
aiogram
pydantic-settings
httpx
pandas
requests
pillow
temp
Python-IO